+++ lib/dice_roller.dart
@@ -1,0 +1,45 @@
+import 'dart:math';
+import 'package:flutter/material.dart';
+
+class DiceRoller extends StatefulWidget {
+  const DiceRoller({super.key});
+  @override
+  State<DiceRoller> createState() {
+    return _DiceRollerState();
+  }
+}
+
+class _DiceRollerState extends State<DiceRoller> {
+  var currentDiceRoll = 2;
+  void rollDice() {
+    setState(() {
+      currentDiceRoll = Random().nextInt(6) + 1;
+    });
+  }
+
+  @override
+  Widget build(context) {
+    return Column(
+      mainAxisSize: MainAxisSize.min,
+      children: [
+        Image.asset(
+          'assets/images/dice-$currentDiceRoll.png',
+          width: 200,
+        ),
+        const SizedBox(
+          height: 20,
+        ),
+        TextButton(
+            onPressed: rollDice,
+            style: TextButton.styleFrom(
+                // padding: const EdgeInsets.only(
+                //   top: 20,
+                // ),
+                backgroundColor: const Color.fromARGB(155, 255, 0, 128),
+                foregroundColor: const Color.fromARGB(255, 255, 255, 255),
+                textStyle: const TextStyle(fontSize: 28)),
+            child: const Text('Roll Dice'))
+      ],
+    );
+  }
+}
