+++ lib/gradient_container.dart
@@ -1,17 +1,13 @@
 import 'package:flutter/material.dart';
+import 'package:trial_app/dice_roller.dart';
 
 class GradientContainer extends StatelessWidget {
-  GradientContainer(this.color1, this.color2, {super.key});
-  GradientContainer.purple({super.key})
+  const GradientContainer(this.color1, this.color2, {super.key});
+  const GradientContainer.purple({super.key})
       : color1 = Colors.deepOrange,
         color2 = Colors.indigo;
   final Color color1;
   final Color color2;
-  var activeDiceImage = 'assets/images/dice-2.png';
-
-  void rollDice() {
-    activeDiceImage = 'assets/images/dice-4.png';
-  }
 
   @override
   Widget build(BuildContext context) {
@@ -23,28 +19,8 @@ class GradientContainer extends StatelessWidget {
           end: Alignment.bottomRight,
         ),
       ),
-      child: Center(
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            Image.asset(
-              activeDiceImage,
-              width: 200,
-            ),
-            const SizedBox(
-              height: 20,
-            ),
-            TextButton(
-                onPressed: rollDice,
-                style: TextButton.styleFrom(
-                    // padding: const EdgeInsets.only(
-                    //   top: 20,
-                    // ),
-                    foregroundColor: const Color.fromARGB(255, 255, 255, 255),
-                    textStyle: const TextStyle(fontSize: 28)),
-                child: const Text('Roll Dice'))
-          ],
-        ),
+      child: const Center(
+        child: DiceRoller(),
       ),
     );
   }
